Convert faults to return an Exception subclass
Piston's rc from which fault is modeled doesn't do that; however, we
prefer doing things the Django (rather than the Piston) way and make
faults be an Exception subclass.

Piston may not do that but it actually gives us all the tools to do it,
with its HttpStatusCode Exception subclass that is catched where
appropriate.

This will allow, for example, for a get_instance_or_404() method to be
written.