Add our own XML emitter
OpenStack's API XML isn't merely an XML representation of the data
structures and therefore Piston's XML Emitter can't be used as-is.

The XML generation code from OpenStack's Nova is guranteed(?) to be
compatible, so re-use their code and adapt it into an Emitter subclass.

Additionally, enable our own Resource implementation that also checks
the Accept header; it was there before but was dead code due to the
missing Emitter.

Finally, make sure that we use "application/xml" instead of "text/xml",
as per the specification