Make ZMQ thread in db_controller stoppable.
Make db_controller run all 0mq operations in a separate StoppableThread
with a stop() method. The main thread calls it, then publishes on an in-proc
PUB socket to interrupt any zmq_recv() call and join the thread when cleaning
up on exceptions/errors.

Minor fix: use the hostname as a prefix for the 0mq SUB socket's identity, to
prevent bugs due to having multiple db_controllers running at the same time on
different machines.