Be consistent about the use of simplejson
simplejson was merged into Python 2.6's stdlib as json. To help with
that and avoid external dependencies, Django provides a utils.simplejson
module that either imports Python's json module or imports an embedded
copy of simplejson.

This was used inconsistenly through the tree, with users importing
either that, or simplejson directly. This fixes all the call sites to
use Django's version.

A notable exception is util/rapi.py that was fixed to import
simplejson/json conditionally without the Django helper so that the
module can be upstream-Ganeti-friendly.