Change console req/response format, add unit tests
Change the format of JSON requests and replies for getting
a server console over VNC. Fix the cloud command-line tool to
understand the new format.

Add unit tests for api/actions.py:get_console().

Add a quick-n-dirty method of detecting if we're running unit
tests in settings.py.dist and set TEST accordingly. If TEST is
used, the API implementation sets dry_run when creating servers
in the backend and assumes mock replies from it.