Add initial VNC console functionality in UI
Add initial support for VNC consoles in the UI.
Add a show_vnc_console() function in JS, which performs a "console"
API action to receive a VNC console endpoint, then connects using
the TightVNC Java Viewer in a popup window.

This is an initial buggy implementation, see #391 for what remains
to be done.

Refs #391