Remove RAPI GetInstanceConsole() call
Remove RAPI GetInstanceConsole() call, because it returns a
potentially unusable endpoint based on cluster-wide vnc_bind_address,
usually 0.0.0.0 or 127.0.0.1. Until this is fixed on the Ganeti side,
we construct a console info reply based on the primary node and network
port as reported by RAPI GetInstance(), fixes #783.

Allow exceptions to propagate when talking to the VNC authentication
proxy, fixes #787.