Agree on consistent instance tags for firewalling
Update both the default API settings and the kvm-vif-bridge script
to use consistent {protected, unprotected, limited} tags to determine
the firewall profile to use on public network interfaces.

The API spec still defines the {ENABLED, DISABLED, PROTECTED} keywords,
which are used between the UI and the API and will updated at a later stage.