snf-ganeti-eventd init script: saner master checks
In snf-ganeti-eventd init script:
Introduce check_ganeti_master() to discover if we are running on the
ganeti master. check_ganeti_master is called during start. It is also
called during restart, after the daemon has been stopped.

We purposely don't check during stop, because we want to succeed even
if ganeti is corrupt (e.g. due to configuration data corruption).