Amend INSTALL_REQUIRES with explicit versions
Amend INSTALL_REQUIRES in */setup.py with explicit version numbers
where not previously specified.

Use '>=' comparisons, to avoid console scripts barfing if the
exact same version specified in setup.py is not currently installed
in the system.

dh_python2 does not seem to pass version requirements in the final
deb produced, this is an open bug in Debian:

http://bugs.debian.org/cgi-bin/bugreport.cgi?bug=653740