Update modules interacting with AMQP
Update all modules which are interacting with the RabbitMQ messaging system to
use the AMQPClient in order to use highly available queues in active/active
mode.

Extend messages to include a unique identifier in order to handle
duplication of messages and out of order delivery. The timestamp of the event
that the message is describing is chosen for this.

Extend VirtualMachine model with 'backendtime' field to hold the
timestamp described above.