Console popup window fix for safari, Refs #2463
It seems that Safari allows popup windows only if the window.open call
is made from within an html element click event context. Otherwise its
behaviour is based on "Block Pop-Up Windows" setting, which when enabled
no notification appears for the user that a popup window was requested
to open.

Since there is no easy way to check the setting value, the action call
changed to use synchronous request. That way the success handler which
calls the window.open method, gets executed within the confirm button
click event context.

In contradictory to this change the user might experience unresponsive
behaviour of the ui when console actions too long to repsond.

The behaviour is limited only for Safari clients.