Encrypt Backend password
Use cipher to encrypt the passwords of the Ganeti backends in the DB.
Encryption/decryption is based on a configurable key, defined by the
ENCRYPTION_KEY setting. After modifying this setting, synnefo will be
unable to decrypt already stored passwords, and passwords must be
re-entered, using 'snf-manage backend-modify' command and restaring the
server. A migration file is responsible for initially encrypting already
stored passwords.