Implement MAC Prefix Management
In order to guarantee uniqueness of MAC address accross all backends,
synnefo assignes a unique mac prefix to each network in each backend.

A mac-prefix of 20 bits, which is picked from the MacPrefixPool, is
assigned to each Network. The mac-prefix of the network in each
backend results from combining the mac-prefix of the Network with the
unique index of the Backend, resulting in a 3-byte mac-prefix (up to 16
Backends).

Settings PRIVATE_MAC_FILTERED_BASE_MAC_PREFIX and
PRIVATE_MAC_FILTERED_MAX_PREFIX_NUMBER have been converted to
MAC_POOL_BASE and MAC_POOL_LIMIT, since mac-prefixes are now assigned to
all networks.