Implement multiple public networks
Implement multiple public networks in order to support:
- different public networks per backend
- non-consecutive IP ranges in each backend

To support this, public networks are treated different from private
ones. Namely, they are correlated with one backend (and so have
only one BackendNetwork entry) instead of private which are expanded
to all available backends.

Also the backend allocator is modified in order to exclude the backends
that have only full networks.

Finally GANETI_PUBLIC_NETWORK became obsolete, and so removed.

Currently, choice of public network is done by synnefo, but in the
future the user may be able to choose it (needs extension of the
api with network-id)