Refactor bridge and mac-prefix pool tables
Redesign the implementation of pools and map mac-prefix and bridge pools
to that design.

Each pool type is mapped to a table in the database, and each row of
that table is a pool of that type. All pool tables must contain the
following columns:

- size: the size of the pool
- available_map: mapping of the available values to a string
- reserved_map: mapping of the external reservations to a string

Access to each pool is managed by a class subclassing the PoolManager
class, and must implement 'index_to_value' and 'value_to_index' methods
which allow mapping of the pool values to an index on a bitarray.

Also the corresponding migration files are created.