Refactor IP pools
Treat IP pool the same way as bridge and mac-prefix pools. The pool is
removed from the Network table and moved to an IPPoolTable containing the
IP pools for all networks. Also added support for external reservations.

Migration files for transfering the pool to the new tables are included.
Finally, the map is inverted, since the new pool treats different the
availability of the values in it.