Implement queue-retry management command
Implement a new command which consumes all the messages from the
dead-letter queues and re-sends the messages back to the snf-dispatcher
in order to handle them again.

Messages are annotated with "from-dead-letter" field in order to detect
looping between the original exchange and the dead-letter exchange. If a
loop is detect the message is discarder, unless the --keep-zombies
option is used.