Create IPPool when creating Network
Create the IPPool associated with each Network, when creating it. Until
now, IPPools were created on first demand. However, because the Network
object is not accessed in exclusive mode, concurrent access to the
Network model resulted in a race condition. More than one workers may
find that no Pool is associated with the Network, and try to create a
new one.