Detect and handle connection problems in amqp_puka
Use the TCP keepalive feature in amqp_puka in order to detect connection
errors to the RabbitMQ broker, and retry to a different one. Also setup
the keepalive parameters to relative small values, to detect failures
relatively soon.

Also, modify and-dispatcher to use the AMQP prefetch_count option of the
basic_consume method, which makes the RabbitMQ broker to not deliver
more than prefetch_count messages to the snf-dispatcher, before
receiving acknowledgement for the previous messages.