Reconnect to different Broker if idle for too long
Modify snf-dispatcher and add a timeout option to basic_wait method.  If
timeout is reached, the connection to the current RabbitMQ broker is
closed and dispatcher tries to connect to a different one.

This commit is supplementary of 6d27ead, which detects connection
failures using the TCP keepalive feature. However, while the connection
is active, the RabbitMQ broker may have a problem. This commit tries
to detect such problems, and considers idle connections such a case.