Issue commissions when creating server/networks
Issue commissions to the Quotaholder when creating a VM or a Network.
The VM/Network that is associated with the corresponding serial in DB.
Also, the serial is marked as accepted because at the end of the
transaction the resource will be present in the DB. Even in case the
resource will not be successfully created in the Ganeti, the user has to
explicitly delete it, in order to deallocate the resources.