Add autopkg.conf and some fixes
The script sources the conf file in order to get build options.

In case building procedure fails, it cleans up the mess and returns
HEAD to initial state.

If it succeeds, it checks if your local branch is up-to-date with the
corresponding upstream. If yes, it let you in temp debian branch
(tagged if in Release mode) and lets you do any push (if desired)
and cleanup (tag and branch removal).

Signed-off-by: Dimitris Aragiorgis <dimara@grnet.gr>