allow overlimit entities to release resources
When a resource holding is found with quantity exceeding its capacity,
it cannot allocate any more quantities,
but it should be able to release quantities.
However, if even after the release the quantity remained
above the capacity, the release was denied with NoCapacity error.

This was fixed by disabling this NoCapacity check when releasing
(i.e. allocated quantity  is negative)