fix+test for issue_commission release logic
issue_commission() with negative quantity is considered
to be a 'release' instead of an 'allocation'.
The limit checks for the conditions or the release were
somewhat incomplete and inconsistent.

The new logic also fixes the bug found by kpap where
an entity would be denied release of quantities (with NoCapacityError)
if the remaining quantity exceeds the quota limit.
This prevented an entity from releasing its overlimit quantities
in multiple steps (e.g. one by one), which resulted in a deadlock,
when a service (cyclades VM) tried to rectify the overlimit cases.