Rollback, even if transaction is not dirty
This is a temporary bug fix for
https://code.djangoproject.com/ticket/9964 .
Django prior to 1.3 does not set a transaction
dirty when the DB throws an exception, and thus
does not trigger rollback, resulting in a
dangling aborted DB transaction.

Since 1.3, django sets a transaction dirty on
attempting any query. This covers the above case,
too.