add capacity to service resource reserves
This is not normally needed, since all resources circulating through
the system must be allocated from the initial reserve quantities.
However, if somehow this consistency is broken, and more quantities
are released to the reserves than have been allocated from them,
the reserves will not be able to accommodate the extra quantities.

Adding (practically infinite) capacity to reserves allows this
inconsistency to be.