Fix transaction handling when creating server
Better handling of transactions when creating a server. Commit the
creation of VM in DB, before enqueuing the job to the Ganeti backend.
If communication with Ganeti fails, then the VM is deleted from the DB.
If the servers crashes before deleting the VM, the stale VM will be
removed by the reconciliation mechanism.