Fix reconciliation for building VMs
Since commit 482c645, a VM exists in the Cyclades DB before the
corresponding job is enqueued in the Ganeti backend. Therefore, there is
a corner case, where Cyclades can not enqueue the job to Ganeti and
Cyclades crashes before deleting the VM from the DB, which results in a
stale VM entry in a building state. This commit adds the necessary check
to guarantee that these VMs will be removed by the reconciliation
process.