Deactivate pithos disk space usage tests
Pithos usage tests require special setup
since test accounts should be created and configured properly
in astakos and quotaholder components.

Also, these tests should know the server's version accounting policy.
For the time being, pithos server does not return this information.
So the tests check the PITHOS_BACKEND_FREE_VERSIONING setting
to discover the server's policy.
However, this works only if the tests run against a server
running in the same host, but not also for a remote server.

For the above reasons, the specific tests are deactivated
until there will be available a proper pithos testing framework.