Preserve owner info when modifying project
An admin can modify a project, i.e. create an application based on
a previous application. We distinguish the initial applicant, called
the `owner', from the user who submitted the current application,
called the `applicant'.

Show owner instead of applicant in project-list: the applicant can
vary among applications for a single project, but owner is likely
to be unique.