Delete Image without deleting file from Pithos
Until now, deleting an image resulted in deletion of the corresponding
file from Pithos. Plankton Images are files in Pithos with special
metadata on the 'plankton' metadata domain. Adding an Image to plankton
is equivalent with adding some special metadata in this domain. In the
same way, this commit implements Image deletion by deleting these
metadata, and retaining the original file in Pithos.

Refs #3522