Allow helpdesk users to shutdown/start vms
Include an additional action to helpdesk view. Based on the current
status of the virtual machine helpdesk user is allowed to start it or
shut it down.

This comes handy since in most cases vm suspension needs to be
followed up with some kind of network cut off of the corresponding
machine. Since no such option is currently available from the
software, machine shutdown is provided as an alternative.