Sync to quotaholder by recomputing user quotas
On a user joining/leaving a project or on a project definition change,
we used to compute quota diffs for affected users and add them up
to existing values in quotaholder.

We now switch to recomputing all quotas for affected users and resetting
the quotaholder accordingly. This has two advantages:

  - we need only one primitive for setting limits (add_quota becomes
    redundant) and
  - we don't need to keep synced/pending pairs of project application on
    project membership; changes are applied instantly.