Import resources from definition and config files
Register a service's resources giving option --json <file>.
The json file is expected to represent a dictionary with keys
"service" and "resources".

Currently the default base quota (uplimit) for each resource is read
from a configuration json file, given with --conf <file>. This file
represents a dictionary with numeric values for each resource name.

Upon importing the resources, the quotaholder has to be updated due to
the modified resource limits.