pithos: Change conflict handling in object write
Pithos in case of missing blocks during smart object write
(using hashmap) used to raise an api fault with payload
the list with the missing blocks serialized in json format.

However, the common api_method decorator in case of api fault
returns a different response content format containing details about
the fault.

This causes problem to pithos clients which expect
a specific response content format in that case.

The solution was adopted is:
when pithos server detects such a case returns directly
an HttpResponse with the same status code (409) as before
and the expected content.