pithos: Fix container quota violation control
Pithos backend before performing the quota violation checks
creates a new version of object and updates the
usage of the ancestor nodes (the container).

In Postgresql SELECT does see the effects
of previous updates executed within its own transaction
even though they are not yet committed.
(regardless of the Transaction Isolation Level)
http://www.postgresql.org/docs/8.4/static/transaction-iso.html

As a result, during the quota checks the backend sees
the updated usage of the container,
so it should not add to this the size difference.