Handle max pending applications as a quotable resource
Assume a resource `astakos.pending_app' related to service `astakos'.
On submitting an application, issue a commission that will update the
related counter if possible, and accept the commission on success.
On approving/denying/cancelling an application, update the counter
likewise.

We always charge the application `owner'; however, if the `applicant'
is a project admin, we issue commission in force mode which succeeds
even if it exceeds the limit.

In order to pre-emptively check in the UI, follow the same process in
`dry run' mode, which issues a commission and rejects it on success.

Remove option of user-update to set its limit.

Refs #3349