Varous fixes to cyclades-usage-verify command
Extend 'cyclades-usage-verify' management command to cover the usage of
'cyclades-reset-usage' and remove the second one. The new command
reconciles resources(usage) by detecting inconsistencies between the
usage of resources in Astakos and Cyclades DB and fixing them by issuing
a commission with the difference of the usage in DB and Astakos. If
there is a pending commission for a resource, reconciliation for this
resource is skipped. Refs #3627

Also, make this command do not consider only users that exist in
Cyclades DB, but also users that exist in Quotaholder. For such users,
it is assumed that they have zero usage for all Cyclades resources.
Refs #3606 #3491