cyclades: refactor code relative to commission
Refactor Cyclades code that handles commissioning with Quotaholder. This
commit removes the 'uses_commission' decorator. Instead it introduces
the 'issue_and_accept_commission' function that is responsible for
implenting the commission workflow. This function  must be called
exactly after and in the same transaction that creates/modifies the
resource, because resource creation/modification must happen in the same
transaction with issuing and storing the serial in the DB.