astakos: Separate importing resources from setting limits
`snf-manage resource-import' does not provide a way to change the
default base quota any more. When importing a resource for the first
time, it sets the limit to zero.

Use `snf-manage resource-modify' to change limits. It supports both
interactive mode and reading limits from file, either for a specified
resource or for all resources.