cyclades: Check qh resolve commission response
Check Quotaholder response to see if the commissions were accepted or
rejected as requested, or failed for some reason.

Accepting or rejecting a commission will fail if the commission was
already resolved. Since this is sometimes accepted, like when
reconciling commissions, this commit extends 'qh_resolve_commissions'
function with the 'strict' argument, denoting whether the function
should fail if resolving a commissions fails.