cyclades: Add FloatingIP model
Add Floating IP model to represent a floating IP that has been reserved
by a user. Each FloatingIP has an owner(UUID), an IPv4 address, the
network(pool) that has reserved from and the date is was reserved. Also,
it may be related to a VirtualMachine, if there is some VM that uses
this IP.

Also, create the corresponding model factory for tests.