astakos: Strengthen authentication in user_from_token
All user-related API calls make use of this decorator to authenticate
the user. Previously it used to check only for the existence of the token;
now it makes all checks previously found in the `authenticate' API call.
The latter now just prepares the response data; it depends on the
decorator for authentication per se.