cyclades: Add network policy at server create
Add 'DEFAULT_INSTANCE_NETWORKS' setting that can be used by the
administrator to define the list of networks that each server that is
created must be connected to. Each created VM will have a NIC connected
to each of these networks. This settings can include a list of network
UUIDs or the special "public" ID, which corresponds to any public
network that has a free IP address.

Also, extend POST /servers API call, with an optional 'networks'
attribute, that the user can use to define a list of private networks
that wants it's VM to be connected.