astakos: Convert list-like settings to tuple
Settings ADMINS, MANAGERS, and HELPDESK are tuples even though they could
naturally be lists. If one sets by mistake one of these to list, this
triggers an error in notification code, which tries to concatenate them:
concatenation of list and tuple fails. To avoid this, we make sure that
these settings are in tuple form.

Also explicitly set MANAGERS and HELPDESK in the default admin settings
to an empty tuple rather than referring to ADMINS (which is an empty
tuple, too).