cyclades: Choose floating-ips when creating server
Extend POST /servers API call, to take an optional attribute
'floating_ips', which is a list of floating IP addresses. The server
will have one NIC for each of this addresses after the
'DEFAULT_INSTANCE_NETWORKS' and before any private networks that the
user has chosen. All of the addresses must first be reserved through the
/os-floating-ips API.