Distinquish between networks and Floating IP pools
Extend Network model with the 'floating_ip_pool' boolean field, to
indicate whether a Network can be a floating IP pool or not. Networks
that are floating IP pools must exist in all Backends, and so
'snf-manage backend-add' should try to add them.

Also, update reconciliation to create networks that are floating IP
pools to the Ganeti backends that do not exist.